Moved gallery out & moved scroll to the inner container
