+++ backend/open-api.json
@@ -107,7 +107,11 @@
             "type": "string"
           }
         },
-        "required": ["timestamp", "originalImgUrl", "source"]
+        "required": [
+          "timestamp",
+          "originalImgUrl",
+          "source"
+        ]
       },
       "ErrorDto": {
         "type": "object",
@@ -120,7 +124,10 @@
             "type": "string"
           }
         },
-        "required": ["statusCode", "message"]
+        "required": [
+          "statusCode",
+          "message"
+        ]
       }
     }
   }
